lib: fwts_battery: use size_t instead of int
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>

+++ src/lib/src/fwts_battery.c
@@ -38,8 +38,8 @@ static int fwts_battery_get_capacity_sys_fs(fwts_framework *fw,
 	struct dirent *entry;
 	char *field_mAh;
 	char *field_mWh;
-	int  field_mAh_len;
-	int  field_mWh_len;
+	size_t field_mAh_len;
+	size_t field_mWh_len;
 	int  i = 0;
 
 	switch (type) {
